+++ packages/lib/src/components/Card/components/CardInput/components/types.ts
@@ -1,7 +1,7 @@
 import { PaymentAmount } from '../../../../../types';
 import { CardBrandsConfiguration } from '../../../types';
 import { ComponentChildren } from 'preact';
-import { CVCPolicyType, DatePolicyType } from '../../../../internal/SecuredFields/lib/securedField/AbstractSecuredField';
+import { CVCPolicyType, DatePolicyType } from '../../../../internal/SecuredFields/lib/types';
 
 export interface BrandIconProps {
     brand: string;
