+++ packages/lib/src/components/Card/components/CardInput/types.ts
@@ -3,7 +3,7 @@ import { CardBrandsConfiguration, CardConfiguration, DualBrandSelectElement, Soc
 import { PaymentAmount } from '../../../../types';
 import { InstallmentOptions } from './components/types';
 import { ValidationResult } from '../../../internal/PersonalDetails/types';
-import { CVCPolicyType, DatePolicyType } from '../../../internal/SecuredFields/lib/securedField/AbstractSecuredField';
+import { CVCPolicyType, DatePolicyType } from '../../../internal/SecuredFields/lib/types';
 import { ValidationRuleResult } from '../../../../utils/Validator/Validator';
 import Specifications from '../../../internal/Address/Specifications';
 import { AddressSchema, StringObject } from '../../../internal/Address/types';
